v0.1: upload the driving map as a media file.
There is limitation on place number when you create driving map based a map tool. Usually the number is 10, 15 or 20. I have just created a driving map as uploaded. It is a way to git rid of that limitation. The procedure was based on Baidu map, Natron and ffmpeg on Linux. The readme will come here soon when I get time.